+++ apps/workflows-dashboard/src/domains/workflow-definitions/workflow-definitions/workflow-definitions.api.ts
@@ -3,6 +3,9 @@ import {
   GetWorkflowDefinitionDto,
   GetWorkflowDefinitionsListDto,
   IWorkflowDefinition,
+  UpdateWorkflowDefinitionByIdDto,
+  UpdateWorkflowDefinitionExtensionsByIdDto,
+  UpgradeWorkflowDefinitionVersionByIdDto,
 } from '@/domains/workflow-definitions/workflow-definitions/workflow-definitions.types';
 import { request } from '@/lib/request';
 
@@ -38,4 +41,30 @@ export const fetchUIDefinitionByWorkflowDefinitionId = async (query: GetUIDefini
   );
 
   return result.data || {};
+};
+
+export const updateWorkflowDefinitionById = async (dto: UpdateWorkflowDefinitionByIdDto) => {
+  const result = await request.put(`/workflow-definition/${dto.workflowDefinitionId}/definition`, {
+    definition: dto.definition,
+  });
+
+  return result.data;
+};
+
+export const updateWorkflowDefinitionExtensionsById = async (
+  dto: UpdateWorkflowDefinitionExtensionsByIdDto,
+) => {
+  const result = await request.put(`/workflow-definition/${dto.workflowDefinitionId}/extensions`, {
+    extensions: dto.extensions,
+  });
+
+  return result.data;
+};
+
+export const upgradeWorkflowDefinitionVersionById = async (
+  dto: UpgradeWorkflowDefinitionVersionByIdDto,
+) => {
+  const result = await request.post(`/workflow-definition/${dto.workflowDefinitionId}/upgrade`);
+
+  return result.data;
 };