+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/config
 
+## 1.1.8
+
+### Patch Changes
+
+- Bump
+
 ## 1.1.7
 
 ### Patch Changes